Add the support for the services
This patch adds the capability to emulate the services in Kubernetes
with Neutron's load balancer resources, pools, VIPs and members.

Each K8s service is mapped into a Neutron pool and a Neutron VIP and
pod containers are treated as Neutorn members. This works well inside
the cluster but we have the following limitations:

* The host can't reach the services because everything is done through
  the OVS datapath
* The K8s services can have multiple port mappings for the single
  cluster IP (VIP in K8s) but Neutron (MidoNet plugin) can handle only
  the pair of the VIP and the port

The problems above shall be resolved in the following patches against
this plugin or Neutorn (MidoNet plugin).

Signed-off-by: Taku Fukushima <tfukushima@midokura.com>